# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Turkcell SkillHub projesi için Backend API dokümantasyonu.",
        "title": "SkillHub API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Destek",
            "url": "http://www.skillhub.com",
            "email": "support@skillhub.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/complete-profile": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Eksik olan şehir, üniversite ve yetenek bilgilerini günceller.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Profil Tamamlama",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Profil Bilgileri",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.UpdateProfileRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Profil güncellendi",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Eksik parametreler",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/secure/me": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Giriş yapmış kullanıcının detaylarını döner.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Profil Bilgilerini Getir",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.UserResponse"
                        }
                    },
                    "401": {
                        "description": "Yetkisiz işlem",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "406": {
                        "description": "Profil eksik (Tamamlanmalı)",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Kullanıcı hesabını silinmiş olarak işaretler.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Hesabı Sil (Soft Delete)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Hesabınız silindi",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Silme başarısız",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/login": {
            "post": {
                "description": "Email ve şifre ile giriş yapar, JWT Token döner.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Giriş Yap",
                "parameters": [
                    {
                        "description": "Giriş Bilgileri",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Geçersiz veri",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Giriş başarısız",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/register": {
            "post": {
                "description": "Sisteme yeni bir kullanıcı ekler.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Yeni Kullanıcı Kaydı",
                "parameters": [
                    {
                        "description": "Kayıt Bilgileri",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.RegisterRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Kayıt başarılı",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Hatalı veri",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "409": {
                        "description": "Bu email zaten kullanımda",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "domain.LoginRequest": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "domain.RegisterRequest": {
            "type": "object",
            "required": [
                "email",
                "name",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string",
                    "minLength": 6
                }
            }
        },
        "domain.UpdateProfileRequest": {
            "type": "object",
            "required": [
                "city",
                "skill_level",
                "university"
            ],
            "properties": {
                "city": {
                    "type": "string"
                },
                "skill_level": {
                    "type": "string"
                },
                "university": {
                    "type": "string"
                }
            }
        },
        "domain.UserResponse": {
            "type": "object",
            "properties": {
                "city": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}