# backend/internal/domain/user.go
package domain

import (
	"time"
	"gorm.io/gorm"
)

// User Entity (DB Tablosu ile eşleşir)
type User struct {
	ID         uint           `gorm:"primaryKey"`
	Name       string         `gorm:"not null"`
	Email      string         `gorm:"unique;not null"`
	Password   string         `gorm:"not null"`
	City       string         `gorm:"size:50"`
	University string         `gorm:"size:100"`
	SkillLevel string         `gorm:"size:20"`
	CreatedAt  time.Time
	DeletedAt  gorm.DeletedAt `gorm:"index"` // Soft Delete Sihri
}

// DTOs (Data Transfer Objects)
type RegisterRequest struct {
	Name     string `json:"name" validate:"required"`
	Email    string `json:"email" validate:"required,email"`
	Password string `json:"password" validate:"required,min=6"`
}

type LoginRequest struct {
	Email    string `json:"email" validate:"required,email"`
	Password string `json:"password" validate:"required"`
}

type UpdateProfileRequest struct {
	City       string `json:"city" validate:"required"`
	University string `json:"university" validate:"required"`
	SkillLevel string `json:"skill_level" validate:"required"`
}

type UserResponse struct {
	ID    uint   `json:"id"`
	Name  string `json:"name"`
	Email string `json:"email"`
	City  string `json:"city"`
}

// Interfaces (SOLID - Interface Segregation & Dependency Inversion)
type UserRepository interface {
	Create(user *User) error
	FindByEmail(email string) (*User, error)
	FindByID(id uint) (*User, error)
	Update(user *User) error
	Delete(id uint) error // Soft Delete yapacak
}

type AuthService interface {
	Register(req *RegisterRequest) error
	Login(req *LoginRequest) (string, error) // Token döner
}

type UserService interface {
	GetProfile(userID uint) (*UserResponse, error)
	UpdateProfile(userID uint, req *UpdateProfileRequest) error
	DeleteUser(userID uint) error
}